moved controlled input to test file and renamed it
